+++ vhost-device-sound/src/args.rs
@@ -1,16 +1,25 @@
 // SPDX-License-Identifier: Apache-2.0 or BSD-3-Clause
 //! An arguments type for the binary interface of this library.
 
+use std::os::fd::RawFd;
 use std::path::PathBuf;
 
-use clap::{Parser, ValueEnum};
+use clap::{ArgGroup, Parser, ValueEnum};
 
 #[derive(Parser, Debug)]
-#[clap(version, about, long_about = None)]
+#[clap(
+    version,
+    about,
+    long_about = None,
+    group(ArgGroup::new("socket group").required(true).args(&["socket", "socket_fd"])),
+)]
 pub struct SoundArgs {
     /// vhost-user Unix domain socket path.
     #[clap(long)]
-    pub socket: PathBuf,
+    pub socket: Option<PathBuf>,
+    /// vhost-user Unix domain socket FD.
+    #[clap(long)]
+    pub socket_fd: Option<RawFd>,
     /// audio backend to be used
     #[clap(long)]
     #[clap(value_enum)]